+++ turtlebot3_applications/CHANGELOG.rst
@@ -2,7 +2,13 @@
 Changelog for package turtlebot3_applications
 ^^^^^^^^^^^^^^^^^^^^^^^^^^^^^^^^^^^^^^^^^^^^^
 
-1.2.0(2025-04-25)
+1.2.1 (2025-06-04)
+------------------
+* Added turtlebot3_yolo_object_detection pkg
+* Added Python script runs the YOLO model on live camera images and publishes the detection results with bounding boxes
+* Contributor: YeonSoo Noh
+
+1.2.0 (2025-04-25)
 ------------------
 * Support for ROS 2 Humble version
 * Added turtlebot3_aruco_tracker pkg